+++ src/app/api/posts/action/comment/delete/route.js
@@ -1,3 +1,4 @@
+import { getTokenData } from "@/helpers/getTokenData";
 import Comment from "@/models/comment";
 import Post from "@/models/post";
 import { connectToDB } from "@/utils/database";
@@ -7,6 +8,12 @@ export const DELETE = async (request, { params }) => {
   try {
     const {commentId} = await request.json();
     await connectToDB();
+    const tokenData = await getTokenData(request);
+    if (!tokenData) {
+      return new Response(JSON.stringify({ error: "Unauthorized access" }), {
+        status: 401,
+      });
+    }
 
     // Find the post that contains the comment
     const foundComment = await Comment.findByIdAndRemove(commentId);
